Refactored io handlers
