+++ tsdat/io/handlers/csv_handler.py
@@ -1,0 +1,12 @@
+from pydantic import Field
+from ..base import FileHandler
+from ..readers import CSVReader
+from ..writers import (
+    CSVWriter,
+)
+
+
+class CSVHandler(FileHandler):
+    extension: str = "csv"
+    reader: CSVReader = Field(default_factory=CSVReader)
+    writer: CSVWriter = Field(default_factory=CSVWriter)
