+++ tsdat/io/handlers/split_netcdf_handler.py
@@ -1,0 +1,12 @@
+from pydantic import Field
+from ..base import FileHandler
+from ..readers import NetCDFReader
+from ..writers import (
+    SplitNetCDFWriter,
+)
+
+
+class SplitNetCDFHandler(FileHandler):
+    extension: str = "nc"
+    reader: NetCDFReader = Field(default_factory=NetCDFReader)
+    writer: SplitNetCDFWriter = Field(default_factory=SplitNetCDFWriter)
